fix: handle user export errors gracefully (#63)
The user export fails if a component like LDAP provider is
registered. In that case, Keycloak tries to connect to LDAP
to request user info. However, if this connection can't
be established, the user export fails and the complete export
process itself stops.

This fixes the user export by catching potential errors. It logs
the error and continues with the remaining process.

Closes #63