fix(secrets): fix export for clients with overlapping naming and specialized templates (#88)
If two clients share the same name, e.g. eam and eam-ui, and a
specialized template is provided for eam-ui, then, depending on
the order the clients are processed, the eam client is overwriting
the exported file for the eam-ui client.

Fix by sorting the client based on their name. Like this, clients
with short names are processed before. Clients with longer names
will overwrite the previous files then.